Fixed lepton api fetching
