+++ package.json
@@ -46,6 +46,7 @@
         "@trpc/client": "11.0.0-rc.502",
         "@trpc/react-query": "11.0.0-rc.502",
         "@trpc/server": "11.0.0-rc.502",
+        "axios": "^1.7.9",
         "class-variance-authority": "^0.7.0",
         "clsx": "^2.1.0",
         "cmdk": "^0.2.1",
