+++ src/app/admin/components/bookinglist-by-state.tsx
@@ -1,13 +1,17 @@
 import BookingList from './booking-list/booking-list';
 import { api } from '@/trpc/server';
+import { type ReservationState } from '@prisma/client';
 
-export default async function ApprovedPage() {
+interface BookingListProps {
+    state: ReservationState;
+}
+
+export default async function BookingListByState({ state }: BookingListProps) {
     const items = await api.reservation.getReservations({
-        state: 'APPROVED',
+        state,
         limit: 20,
         direction: 'forward',
         cursor: null,
     });
-
     return <BookingList items={items.reservations} />;
 }
