+++ src/server/api/urls.ts
@@ -1,0 +1,3 @@
+export const URLS = {
+    USERS: 'users',
+};
