--- src/types/User.ts
@@ -1,17 +1,0 @@
-export type User = {
-    allergy: string;
-    email: string;
-    first_name: string;
-    gender: number;
-    image: string;
-    last_name: string;
-    tool: string;
-    unread_notifications: number;
-    user_id: string;
-    unanswered_evaluations_count: number;
-    number_of_strikes: number;
-    public_event_registrations: boolean;
-    accepts_event_rules: boolean;
-    allows_photo_by_default: boolean;
-    slack_user_id: string;
-};
